This is a big commit that touches multiple issues.
first of all I have made some changes to Util.pm to
include a StripJavascript function that provides basic
protection against cross site scripting attacks

This commit also includes the "Oyster" plugin architecture.
These are multilayer plugins almost like the pearls but
with a focus on manipulating data instead of reporting.

This release also includes the 'case insensitve' search
fix for the jsisearch tool.

Unfortunately I had to extend the functionality of the
gedafed program to make it a secure storage of temporary data.
I am sure however that we can use this extention for more
purposes.


git-svn-id: svn://svn.ee.ethz.ch/gedafe/trunk/gedafe@367 c5e4ef7b-cb2b-0410-bad8-cbb44b149fe0